+++ analysis_options.yaml
@@ -3,4 +3,6 @@ include: package:carapacik_lints/core.yaml
 linter:
   rules:
     always_use_package_imports: false
+    depend_on_referenced_packages: false
     prefer_relative_imports: true
+    unnecessary_library_name: false
