+++ example/README.md
@@ -1,1 +1,1 @@
-# A project that showcases usage of model_viewer
+# A project that showcases usage of model_viewer_plus
