+++ example/analysis_options.yaml
@@ -1,1 +1,5 @@
 include: package:carapacik_lints/core.yaml
+
+linter:
+  rules:
+    implementation_imports: false
