--- example/android/app/src/main/kotlin/com/carapacik/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.carapacik.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
