+++ example/android/app/src/main/kotlin/com/example/model_viewer_plus/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.model_viewer_plus
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()
