+++ example/ios/Runner/AppDelegate.swift
@@ -1,5 +1,5 @@
-import UIKit
 import Flutter
+import UIKit
 
 @UIApplicationMain
 @objc class AppDelegate: FlutterAppDelegate {
