+++ example/lib/materials_and_scene/change_color.dart
@@ -1,9 +1,8 @@
-// ignore_for_file: implementation_imports
 import 'package:flutter/material.dart';
 import 'package:model_viewer_plus/model_viewer_plus.dart';
 import 'package:model_viewer_plus/src/model_viewer_plus_web.dart';
-import 'package:model_viewer_plus/src/shim/dart_html_fake.dart'
-    if (dart.library.html) 'dart:html';
+import 'package:model_viewer_plus/src/shim/dart_web_fake.dart'
+    if (dart.library.js_interop) 'dart:html';
 
 void main() => runApp(const MyApp());
