+++ example/pubspec.yaml
@@ -1,19 +1,19 @@
 name: example
-description: A project that showcases usage of model_viewer
+description: "A project that showcases usage of model_viewer_plus"
 publish_to: 'none'
 version: 1.0.0+1
 environment:
-  sdk: ^3.1.0
+  sdk: '>=3.4.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-
-dev_dependencies:
-  carapacik_lints: ^1.5.0
   model_viewer_plus:
     path:
       ../
+
+dev_dependencies:
+  carapacik_lints: ^1.8.1
 
 flutter:
   uses-material-design: true
