--- lib/src/shim/dart_ui_fake.dart
@@ -1,11 +1,0 @@
-final PlatformViewRegistry platformViewRegistry = PlatformViewRegistry();
-
-class PlatformViewRegistry {
-  bool registerViewFactory(
-    String viewType,
-    Function viewFactory, {
-    bool isVisible = true,
-  }) {
-    return false;
-  }
-}
