+++ lib/src/shim/dart_web_fake.dart
