+++ pubspec.yaml
@@ -1,30 +1,30 @@
 name: model_viewer_plus
-description: A Flutter widget for rendering interactive 3D models in the glTF and GLB formats
-version: 1.7.1
+description: "A Flutter widget for rendering interactive 3D models in the glTF and GLB formats"
+version: 1.8.0
 repository: https://github.com/omchiii/model_viewer_plus.dart
 topics:
   - model-viewer
   - model
+  - 3d
 platforms:
   android:
   ios:
   web:
 environment:
-  sdk: ^3.1.0
-  flutter: '>=3.13.0'
+  sdk: '>=3.4.0 <4.0.0'
+  flutter: '>=3.22.0'
 
 dependencies:
-  android_intent_plus: ^5.0.1
+  android_intent_plus: ^5.0.2
   flutter:
     sdk: flutter
-  path: ^1.8.3
-  url_launcher: ^6.2.1
-  webview_flutter: ^4.5.0
-  webview_flutter_android: ^3.13.0
-  webview_flutter_wkwebview: ^3.10.0
+  path: ^1.9.0
+  url_launcher: ^6.2.6
+  web: ^0.5.1
+  webview_flutter: ^4.8.0
 
 dev_dependencies:
-  carapacik_lints: ^1.5.1
+  carapacik_lints: ^1.8.1
   flutter_test:
     sdk: flutter
